+++ DocsSnippets/05-Messages/04-Silent-Messages/00-Silent-Messages.swift
@@ -1,5 +1,0 @@
-// LINK: https://getstream.io/chat/docs/ios-swift/silent_messages/?language=swift&preview=1
-
-func snippet_messages_silent_messages() {
-    // coming soon
-}