+++ StreamChatUITestsAppUITests/Robots/UserRobot+Asserts.swift
@@ -4,7 +4,8 @@
 
 import Foundation
 import XCTest
-import StreamChat
+@testable import StreamChatUI
+@testable import StreamChat
 
 let channelAttributes = ChannelListPage.Attributes.self
 let channelCells = ChannelListPage.cells
@@ -95,7 +96,45 @@ extension UserRobot {
         }
         return self
     }
-
+    
+    @discardableResult
+    func assertChannelListPagination(
+        channelsCount expectedCount: Int,
+        file: StaticString = #filePath,
+        line: UInt = #line
+    ) -> Self {
+        ChannelListPage.cells.firstMatch.wait()
+        let expectedChannel = ChannelListPage.channel(withName: "\(expectedCount)")
+        var expectedChannelExist = expectedChannel.exists
+        
+        XCTAssertFalse(expectedChannelExist,
+                       "Expected channel should not be visible",
+                       file: file,
+                       line: line)
+        
+        let endTime = Date().timeIntervalSince1970 * 1000 + XCUIElement.waitTimeout * 1000
+        while !expectedChannelExist && endTime > Date().timeIntervalSince1970 * 1000 {
+            ChannelListPage.list.swipeUp()
+            expectedChannelExist = expectedChannel.exists
+        }
+        
+        XCTAssertTrue(expectedChannelExist,
+                      "Expected channel should be visible",
+                      file: file,
+                      line: line)
+        return self
+    }
+    
+    @discardableResult
+    func assertChannelCount(
+        _ expectedCount: Int,
+        file: StaticString = #filePath,
+        line: UInt = #line
+    ) -> Self {
+        let actualCount = ChannelListPage.cells.waitCount(expectedCount).count
+        XCTAssertEqual(expectedCount, actualCount, file: file, line: line)
+        return self
+    }
 }
 
 // MARK: Message List
@@ -190,6 +229,20 @@ extension UserRobot {
     }
 
     @discardableResult
+    func assertHardDeletedMessage(
+        at messageCellIndex: Int = 0,
+        file: StaticString = #filePath,
+        line: UInt = #line
+    ) -> Self {
+        let messageCell = messageCell(withIndex: messageCellIndex, file: file, line: line)
+        let message = attributes.text(in: messageCell).wait()
+        let expectedMessage = attributes.deletedMessagePlaceholder
+        let actualMessage = message.waitForText(expectedMessage).text
+        XCTAssertEqual(expectedMessage, actualMessage, "Text is wrong", file: file, line: line)
+        return self
+    }
+
+    @discardableResult
     func assertMessageAuthor(
         _ author: String,
         at messageCellIndex: Int? = nil,
@@ -209,8 +262,9 @@ extension UserRobot {
         file: StaticString = #filePath,
         line: UInt = #line
     ) -> Self {
-        let typingIndicatorView = MessageListPage.typingIndicator
-        let typingUserText = typingIndicatorView.waitForText(typingUserName).text
+        let typingIndicatorView = MessageListPage.typingIndicator.wait()
+        XCTAssertTrue(typingIndicatorView.exists, "Element hidden", file: file, line: line)
+        let typingUserText = typingIndicatorView.waitForText(typingUserName, timeout: 10).text
         XCTAssert(typingUserText.contains(typingUserName), file: file, line: line)
         return self
     }
@@ -243,7 +297,7 @@ extension UserRobot {
     ) -> Self {
         let messageCell = messageCell(withIndex: messageCellIndex, file: file, line: line)
         let errorButton = attributes.errorButton(in: messageCell).wait()
-        XCTAssertTrue(errorButton.wait().exists, file: file, line: line)
+        XCTAssertTrue(errorButton.wait(timeout: 10).exists, file: file, line: line)
         return self
     }
 
@@ -259,7 +313,7 @@ extension UserRobot {
         if deliveryStatus == .failed || deliveryStatus == nil {
             XCTAssertFalse(checkmark.exists, file: file, line: line)
         } else {
-            XCTAssertTrue(checkmark.wait().exists, file: file, line: line)
+            XCTAssertTrue(checkmark.wait(timeout: 10).exists, file: file, line: line)
         }
 
         return self
@@ -277,7 +331,7 @@ extension UserRobot {
         if readBy == 0 {
             XCTAssertFalse(readByCount.isHittable, file: file, line: line)
         } else {
-            let actualText = readByCount.waitForText("\(readBy)").text
+            let actualText = readByCount.waitForText("\(readBy)", timeout: 10).text
             XCTAssertEqual("\(readBy)", actualText, file: file, line: line)
         }
         return self
@@ -322,18 +376,100 @@ extension UserRobot {
                                               line: UInt = #line) {
         let messageCell = messageCell(withIndex: messageCellIndex, file: file, line: line)
         let cellHeight = messageCell.height
-        let textView = MessageListPage.Attributes.text(in: messageCell)
+        let textView = attributes.text(in: messageCell)
         let newLine = "new line"
-        let newText = linesCountShouldBeIncreased ? "\(textView.text)\n\(newLine)" : newLine
+        let newText = linesCountShouldBeIncreased ? "ok\n\(textView.text)\n\(newLine)" : newLine
         
         editMessage(newText, messageCellIndex: messageCellIndex)
-        assertMessage(newText, at: messageCellIndex, file: file, line: line)
+        if ProcessInfo().operatingSystemVersion.majorVersion > 12 {
+            // XCUITest does not get text from a cell after editing it on iOS 12
+            assertMessage(newText, at: messageCellIndex, file: file, line: line)
+        }
         
         if linesCountShouldBeIncreased {
             XCTAssertLessThan(cellHeight, messageCell.height, file: file, line: line)
         } else {
             XCTAssertGreaterThan(cellHeight, messageCell.height, file: file, line: line)
         }
+    }
+    
+    @discardableResult
+    func assertMessageListPagination(
+        messagesCount expectedCount: Int,
+        file: StaticString = #filePath,
+        line: UInt = #line
+    ) -> Self {
+        let endTime = Date().timeIntervalSince1970 * 1000 + XCUIElement.waitTimeout * 1000
+        var actualCount = MessageListPage.cells.count
+        XCTAssertNotEqual(expectedCount, actualCount, file: file, line: line)
+        
+        while actualCount != expectedCount && endTime > Date().timeIntervalSince1970 * 1000 {
+            MessageListPage.list.swipeDown()
+            actualCount = MessageListPage.cells.count
+        }
+        
+        XCTAssertEqual(expectedCount, actualCount, file: file, line: line)
+        return self
+    }
+    
+    @discardableResult
+    func assertComposerLeftButtons(
+        shouldBeVisible: Bool,
+        file: StaticString = #filePath,
+        line: UInt = #line
+    ) -> Self {
+        let composer = MessageListPage.Composer.self
+        let leftButtonsVisible = shouldBeVisible
+            ? composer.attachmentButton.wait().exists && composer.commandButton.wait().exists
+            : composer.attachmentButton.waitForDisappearance().exists && composer.commandButton.waitForDisappearance().exists
+        XCTAssertEqual(shouldBeVisible,
+                       leftButtonsVisible,
+                       "Composer left buttons should be visible: \(shouldBeVisible)",
+                       file: file,
+                       line: line)
+        return self
+    }
+    
+    @discardableResult
+    func assertComposerMentions(
+        shouldBeVisible: Bool,
+        file: StaticString = #filePath,
+        line: UInt = #line
+    ) -> Self {
+        let mentionsView = MessageListPage.ComposerMentions.self
+        if shouldBeVisible {
+            let count = mentionsView.cells.waitCount(1).count
+            XCTAssertGreaterThan(count, 0, file: file, line: line)
+        } else {
+            mentionsView.cells.firstMatch.waitForDisappearance()
+            XCTAssertEqual(mentionsView.cells.count, 0, file: file, line: line)
+        }
+        return self
+    }
+    
+    @discardableResult
+    func assertMentionWasApplied(file: StaticString = #filePath, line: UInt = #line) -> Self {
+        let expectedText = "@\(UserDetails.hanSoloName)"
+        let actualText = MessageListPage.Composer.textView.waitForText(expectedText).text
+        XCTAssertEqual(expectedText, actualText, file: file, line: line)
+        return self
+    }
+
+    @discardableResult
+    func assertComposerCommands(
+        shouldBeVisible: Bool,
+        file: StaticString = #filePath,
+        line: UInt = #line
+    ) -> Self {
+        let commandsView = MessageListPage.ComposerCommands.self
+        if shouldBeVisible {
+            let count = commandsView.cells.waitCount(1).count
+            XCTAssertGreaterThan(count, 0, file: file, line: line)
+        } else {
+            commandsView.cells.firstMatch.waitForDisappearance()
+            XCTAssertEqual(commandsView.cells.count, 0, file: file, line: line)
+        }
+        return self
     }
 }
 
@@ -412,19 +548,33 @@ extension UserRobot {
 
     @discardableResult
     func assertCooldownIsShown(file: StaticString = #filePath, line: UInt = #line) -> Self {
-        XCTAssertTrue(MessageListPage.Composer.cooldown.wait().exists)
+        XCTAssertEqual(MessageListPage.Composer.placeholder.text,
+                       L10n.Composer.Placeholder.slowMode,
+                       file: file,
+                       line: line)
+        XCTAssertTrue(MessageListPage.Composer.cooldown.wait().exists,
+                      "Cooldown should be visible",
+                      file: file,
+                      line: line)
         return self
     }
     
     @discardableResult
     func assertCooldownIsNotShown(file: StaticString = #filePath, line: UInt = #line) -> Self {
-        XCTAssertFalse(MessageListPage.Composer.cooldown.exists)
+        XCTAssertNotEqual(MessageListPage.Composer.placeholder.text,
+                          L10n.Composer.Placeholder.slowMode,
+                          file: file,
+                          line: line)
+        XCTAssertFalse(MessageListPage.Composer.cooldown.exists,
+                       "Cooldown should not be visible",
+                       file: file,
+                       line: line)
         return self
     }
 
     @discardableResult
     func assertSendButtonIsNotShown(file: StaticString = #filePath, line: UInt = #line) -> Self {
-        XCTAssertFalse(MessageListPage.Composer.sendButton.waitForLoss().exists)
+        XCTAssertFalse(MessageListPage.Composer.sendButton.waitForDisappearance().exists)
         return self
     }
 }
@@ -441,7 +591,7 @@ extension UserRobot {
         let messageCell = messageCell(withIndex: messageCellIndex, file: file, line: line)
         let reaction = attributes.reactionButton(in: messageCell)
         let errMessage = isPresent ? "There are no reactions" : "Reaction is presented"
-        _ = isPresent ? reaction.wait() : reaction.waitForLoss()
+        _ = isPresent ? reaction.wait() : reaction.waitForDisappearance()
         XCTAssertEqual(isPresent, reaction.exists, errMessage, file: file, line: line)
         return self
     }
@@ -470,12 +620,20 @@ extension UserRobot {
         line: UInt = #line
     ) -> Self {
         let cell = messageCell(withIndex: messageCellIndex, file: file, line: line).wait()
-        XCTAssertTrue(attributes.giphyImageView(in: cell).wait().exists)
         XCTAssertTrue(attributes.giphyLabel(in: cell).wait().exists)
-        XCTAssertTrue(attributes.giphyBadge(in: cell).wait().exists)
-        XCTAssertFalse(attributes.giphySendButton(in: cell).exists)
-        XCTAssertFalse(attributes.giphyShuffleButton(in: cell).exists)
-        XCTAssertFalse(attributes.giphyCancelButton(in: cell).exists)
+        XCTAssertEqual(0, attributes.giphyButtons(in: cell).count)
+        return self
+    }
+
+    @discardableResult
+    func assertGiphyImageNotVisible(
+        at messageCellIndex: Int? = nil,
+        file: StaticString = #filePath,
+        line: UInt = #line
+    ) -> Self {
+        let cell = messageCell(withIndex: messageCellIndex, file: file, line: line)
+        XCTAssertFalse(attributes.giphyLabel(in: cell).waitForDisappearance().exists)
+        XCTAssertEqual(0, attributes.giphyButtons(in: cell).count)
         return self
     }
 
@@ -519,17 +677,17 @@ extension UserRobot {
     
     @discardableResult
     func assertImage(
-        count: Int = 1,
         isPresent: Bool,
         at messageCellIndex: Int? = nil,
         file: StaticString = #filePath,
         line: UInt = #line
     ) -> Self {
-        let messageCell = messageCell(withIndex: messageCellIndex, file: file, line: line)
-        let images = attributes.images(in: messageCell)
-        let errMessage = isPresent ? "There are no images" : "Image are presented"
-        _ = isPresent ? images.firstMatch.wait() : images.firstMatch.waitForLoss()
-        XCTAssertEqual(images.count, count, errMessage, file: file, line: line)
+        let messageCell = messageCell(withIndex: messageCellIndex)
+        MessageListPage.Attributes.time(in: messageCell).wait()
+        tapOnMessage(messageCell)
+        let fullscreenImage = attributes.fullscreenImage().wait()
+        let errMessage = isPresent ? "There is no image" : "Image is presented"
+        XCTAssertTrue(fullscreenImage.exists, errMessage, file: file, line: line)
         return self
     }
     
@@ -540,7 +698,9 @@ extension UserRobot {
         file: StaticString = #filePath,
         line: UInt = #line
     ) -> Self {
-        tapOnMessage(at: messageCellIndex)
+        let messageCell = messageCell(withIndex: messageCellIndex)
+        MessageListPage.Attributes.time(in: messageCell).wait()
+        tapOnMessage(messageCell)
         let player = attributes.videoPlayer().wait()
         let errMessage = isPresent ? "There is no video" : "Video is presented"
         XCTAssertTrue(player.exists, errMessage, file: file, line: line)
@@ -556,12 +716,10 @@ extension UserRobot {
         line: UInt = #line
     ) -> Self {
         let messageCell = messageCell(withIndex: messageCellIndex, file: file, line: line)
-        let fileIcons = attributes.fileIcons(in: messageCell)
         let fileNames = attributes.fileNames(in: messageCell)
         let errMessage = isPresent ? "There are no files" : "Files are presented"
-        _ = isPresent ? fileNames.firstMatch.wait() : fileNames.firstMatch.waitForLoss()
+        _ = isPresent ? fileNames.firstMatch.wait() : fileNames.firstMatch.waitForDisappearance()
         XCTAssertEqual(fileNames.count, count, errMessage, file: file, line: line)
-        XCTAssertEqual(fileIcons.count, count, errMessage, file: file, line: line)
         return self
     }
 }